+++ R/extract_taxonomy--parsers.R
@@ -110,6 +110,22 @@ class_from_class <- function(class, class_key, class_regex, class_sep, class_rev
     })
   }
   
+  # Check that the regex matched something
+  not_matched <- vapply(result, function(x) any(is.na(x)), logical(1))
+  if (any(not_matched)) {
+    max_to_display <- 10
+    input_to_display <- class[1:min(c(length(class), max_to_display))]
+    names(input_to_display) <- which(not_matched)[1:min(c(length(class), max_to_display))]
+    error_msg <- paste0('The classification regex "', class_regex,
+                        '" does not match the following ', sum(not_matched),
+                        ' of ', length(not_matched), ' inputs:\n', 
+                        paste0("  ", names(input_to_display), ": ", input_to_display, collapse = "\n"))
+    if (length(class) > max_to_display) {
+      error_msg <- paste(error_msg, "\n   ...")
+    }
+    warning(error_msg)
+  }
+  
   # Name columns in each classification according to the key
   result <- lapply(result, function(x) stats::setNames(x, names(class_key)))