+++ NEWS.md
@@ -1,5 +1,13 @@
 # News 
 
+## Current 
+
+### Bug fixes
+
+* Fixed bug in `primersearch` that cased the wrong primer sequence to be returned when primers match in the reverse direction
+* Fixed a bug in `parse_mothur_summary` where "unclassified" had got changed to "untaxmap" during a search and replace
+
+
 ## metacoder 0.1.2
 
 ### Breaking changes