+++ NEWS.md
@@ -12,6 +12,7 @@
 
 ### Bug fixes
 
+* `extract_taxonomy`: Fixed an error that occured when not all inputs could be classified and sequences were supplied
 * Fixed bug in `primersearch` that cased the wrong primer sequence to be returned when primers match in the reverse direction
 * Fixed a bug in `parse_mothur_summary` where "unclassified" had got changed to "untaxmap" during a search and replace
 * Fixed outdated example code for `extract_taxonomy`