+++ NEWS.md
@@ -18,6 +18,7 @@
 
 ### Bug fixes
 
+* Fixed bug in `heat_tree` that made values not in the input taxmap object not associate with the right taxa. See [this post](https://groups.google.com/d/msgid/metacoder-discussions/c9d8ecc2-1efa-4baf-946e-0f105575da2e%40googlegroups.com).
 * `extract_taxonomy`: Fixed an error that occured when not all inputs could be classified and sequences were supplied
 * Fixed bug in `primersearch` that cased the wrong primer sequence to be returned when primers match in the reverse direction
 * Fixed a bug in `parse_mothur_summary` where "unclassified" had got changed to "untaxmap" during a search and replace