+++ NEWS.md
@@ -1,6 +1,6 @@
 # News 
 
-## Current 
+##  metacoder 0.1.3 
 
 ### Improvements